acagent/pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "acagent"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "bedrock-agentcore>=1.0.3",
    "mcp >= 1.19.0",
    "pyjwt>=2.12.1",
    "pytest >= 7.0.0",
    "pytest-asyncio >= 0.21.0",
    "requests >= 2.32.5",
    "strands-agents>=1.13.0",
    "strands-agents-tools>=0.2.16",
]
